pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Igor Aukhadeev"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
